중간단계
